add .travis.yml & update python3 dependency to 3.3
This patch adds support for the travis-ci.org CI system.

The "C" environment in .travis.yml is used to avoid having PPA
python versions installed in the travis chroot.

python3.3 is needed so that test_paths.py runs

The tests_paths.py will only work with python3.3 as it uses
the u"" prefix